Add files via upload
This file uses the A* algorithm to estimate the best path from one point to another. It takes user input to define the target point and estimates the shortest path from the starting point (24,0). According to the assignment requirements, the video needs to show the car going to two different targets. I figured the code needs to be flexible and allow different coordinates to be entered as the end point of the path.