Add files via upload
A* algorithm and following the shortest path is added to Rob's 'decide_on_action' function.
I noticed that there is an issue with the sensor readings sometimes; if the car rides between two objects it detects  something in front instead on the side. I think this can be adjusted in the function. Perhaps left should be from 0 to 72 degrees, not 54? 

Please test this and make adjustments accordingly.